redo some lib

+++ libft/ft_wordlen.c
@@ -6,7 +6,7 @@
 /*   By: jkarren <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/29 22:54:19 by jkarren           #+#    #+#             */
-/*   Updated: 2019/09/29 23:00:21 by jkarren          ###   ########.fr       */
+/*   Updated: 2020/03/04 14:56:37 by jkarren          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_wordlen(char *str, const char lim_ch)
 {
-	int len;
+	size_t	len;
 
 	len = -1;
 	while (str[++len] && str[len] != lim_ch)
